core_api/requirements.txt
fastapi
uvicorn[standard]
neo4j
python-dotenv
pydantic
